requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-rest-auth==6.0.0
Django==4.2.13
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
idna==3.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
